# Conflicts:
#	base.app.backoffice.console/src/main/java/eapli/base/app/backoffice/console/presentation/MainMenu.java

+++ base.core/src/main/java/eapli/base/clientusermanagement/application/AssociateRemoveCollaboratorTeamController.java
@@ -5,7 +5,7 @@ import eapli.base.clientusermanagement.domain.MecanographicNumber;
 import eapli.base.clientusermanagement.dto.ClientUserDTO;
 import eapli.base.clientusermanagement.repositories.ClientUserRepository;
 import eapli.base.infrastructure.persistence.PersistenceContext;
-import eapli.base.teamManagement.application.TeamService;
+
 import eapli.base.teamManagement.domain.Team;
 import eapli.base.teamManagement.domain.Uniquecode;
 import eapli.base.teamManagement.dto.TeamDTO;
@@ -25,7 +25,7 @@ public class AssociateRemoveCollaboratorTeamController {
     private final ClientUserRepository collaboratorRepository = PersistenceContext.repositories().clientUsers();
     private final TeamRepository teamRepository = PersistenceContext.repositories().team();
     private final ClientUserService collaborators =  new ClientUserService();
-    private final TeamService teams = new TeamService();
+  //  private final ListTeamService teams = new ListTeamService();
 
     public void associateCollaboratorTeamController(String collaboratorID, String teamID) throws IllegalAccessException {
         authz.ensureAuthenticatedUserHasAnyOf(BaseRoles.POWER_USER,
@@ -39,8 +39,8 @@ public class AssociateRemoveCollaboratorTeamController {
                 throw new IllegalArgumentException("Collaborator already belongs to this team type");
             }
             else {
-                collaborator.get().teamList().add(team.get());
-                team.get().collaboratorList().add(collaborator.get());
+                collaborator.get().addTeam(team.get());
+               /* team.get().addCollaborator(collaborator.get());*/
 
                 collaboratorRepository.delete(collaborator.get());
                 collaboratorRepository.save(collaborator.get());
@@ -55,7 +55,13 @@ public class AssociateRemoveCollaboratorTeamController {
 
     }
 
+    public Iterable<ClientUserDTO> collaboratorList(){
+        return collaborators.findAllClientUser();
+    }
 
+   /* public Iterable<TeamDTO> teamListWithoutThisCollaborator(String collaboratorID){
+        return teams.teamListWithoutThisCollaborrator(new MecanographicNumber(collaboratorID));
+    }*/
 
     public void removeCollaboratorTeamController(String collaboratorID, String teamID){
         authz.ensureAuthenticatedUserHasAnyOf(BaseRoles.POWER_USER,
@@ -63,30 +69,24 @@ public class AssociateRemoveCollaboratorTeamController {
 
         Optional<ClientUser> collaborator = collaboratorRepository.findByMecanographicNumber(new MecanographicNumber(collaboratorID));
         Optional<Team> team = teamRepository.ofIdentity(new Uniquecode(teamID));
-
+/*
         if (collaborator.isPresent() && team.isPresent()){
-
-                collaborator.get().teamList().remove(team.get());
-                team.get().collaboratorList().remove(collaborator.get());
+            if (collaborator.get().belongToThisTeamType(team.get())){
+                throw new IllegalArgumentException("Collaborator already belongs to this team type");
+            }
+            else {
+                collaborator.get().addTeam(team.get());
+                team.get().addCollaborator(collaborator.get());
 
                 collaboratorRepository.delete(collaborator.get());
                 collaboratorRepository.save(collaborator.get());
 
                 teamRepository.delete(team.get());
                 teamRepository.save(team.get());
-        }
+            }
+        }*/
     }
-
-
-    public Iterable<ClientUserDTO> collaboratorList(){
-        return collaborators.findAllClientUser();
-    }
-
-    public Iterable<TeamDTO> teamListWithoutThisCollaborator(String collaboratorID){
-        return teams.teamListWithoutThisCollaborrator(new MecanographicNumber(collaboratorID));
-    }
-
-
+/*
     public Iterable<TeamDTO> teamList(){
         return teams.teams();
     }
@@ -98,5 +98,7 @@ public class AssociateRemoveCollaboratorTeamController {
     public Iterable<ClientUserDTO> teamCollaborators(String teamid){
         return teams.teamCollaborators(new Uniquecode(teamid));
     }
+
+ */
 }
